requirements.txt
Pillow>=10.0.0
rumps>=0.4.0
pyobjc-framework-Quartz
pyobjc-framework-Cocoa
pytest>=7.0.0